+++ src/Database/ClientDefinition.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Heptacom\AdminOpenAuth\Database;
 
-use Heptacom\AdminOpenAuth\Database\Aggregate\ClientAclRoleDefinition;
-use Shopware\Core\Framework\Api\Acl\Role\AclRoleDefinition;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityDefinition;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\BoolField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\CreatedAtField;
@@ -14,7 +12,6 @@ use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\PrimaryKey;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\Flag\Required;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\IdField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\JsonField;
-use Shopware\Core\Framework\DataAbstractionLayer\Field\ManyToManyAssociationField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\OneToManyAssociationField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\StringField;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\UpdatedAtField;
@@ -43,7 +40,6 @@ final class ClientDefinition extends EntityDefinition
     {
         return [
             'active' => false,
-            'userBecomeAdmin' => true,
             'keepUserUpdated' => true,
         ];
     }
@@ -58,7 +54,6 @@ final class ClientDefinition extends EntityDefinition
             (new BoolField('login', 'login'))->addFlags(new Required()),
             (new BoolField('connect', 'connect'))->addFlags(new Required()),
             (new BoolField('store_user_token', 'storeUserToken'))->addFlags(new Required()),
-            (new BoolField('user_become_admin', 'userBecomeAdmin'))->addFlags(new Required()),
             (new BoolField('keep_user_updated', 'keepUserUpdated'))->addFlags(new Required()),
             (new JsonField('config', 'config', [], []))->addFlags(new Required()),
             new CreatedAtField(),
@@ -69,8 +64,6 @@ final class ClientDefinition extends EntityDefinition
             (new OneToManyAssociationField('userKeys', UserKeyDefinition::class, 'client_id', 'id'))->addFlags(new CascadeDelete()),
             (new OneToManyAssociationField('userTokens', UserTokenDefinition::class, 'client_id', 'id'))->addFlags(new CascadeDelete()),
             (new OneToManyAssociationField('rules', ClientRuleDefinition::class, 'client_id', 'id'))->addFlags(new CascadeDelete()),
-
-            (new ManyToManyAssociationField('defaultAclRoles', AclRoleDefinition::class, ClientAclRoleDefinition::class, 'client_id', 'acl_role_id'))->addFlags(new CascadeDelete()),
         ]);
     }
 }