+++ src/Http/Route/Support/RedirectReceiveRoute.php
@@ -4,19 +4,13 @@ declare(strict_types=1);
 
 namespace Heptacom\AdminOpenAuth\Http\Route\Support;
 
-use Heptacom\AdminOpenAuth\Contract\Client\ClientContract;
 use Heptacom\AdminOpenAuth\Contract\Client\ClientFactoryContract;
-use Heptacom\AdminOpenAuth\Contract\OAuthRuleScope;
 use Heptacom\AdminOpenAuth\Contract\RedirectBehaviour;
-use Heptacom\AdminOpenAuth\Contract\RoleAssignment;
 use Heptacom\AdminOpenAuth\Contract\Route\Exception\RedirectReceiveException;
 use Heptacom\AdminOpenAuth\Contract\Route\Exception\RedirectReceiveMissingStateException;
 use Heptacom\AdminOpenAuth\Contract\User;
-use Heptacom\AdminOpenAuth\Database\ClientRuleCollection;
-use Heptacom\AdminOpenAuth\Service\ClientRuleValidator;
+use Heptacom\AdminOpenAuth\Database\ClientEntity;
 use Psr\Http\Message\RequestInterface;
-use Psr\Log\LoggerInterface;
-use Shopware\Core\Framework\Context;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 class RedirectReceiveRoute
@@ -24,8 +18,6 @@ class RedirectReceiveRoute
     public function __construct(
         private readonly ClientFactoryContract $clientFactory,
         private readonly EventDispatcherInterface $eventDispatcher,
-        private readonly ClientRuleValidator $clientRuleValidator,
-        private readonly LoggerInterface $logger,
     ) {
     }
 
@@ -34,10 +26,8 @@ class RedirectReceiveRoute
      */
     public function onReceiveRequest(
         RequestInterface $request,
-        string $providerKey,
-        array $configuration,
+        ClientEntity $client,
         RedirectBehaviour $behaviour,
-        ClientRuleCollection $rules,
     ): User {
         \parse_str($request->getUri()->getQuery(), $getParams);
 
@@ -55,46 +45,23 @@ class RedirectReceiveRoute
             throw new RedirectReceiveMissingStateException($params, $behaviour->stateKey);
         }
 
-        $client = $this->clientFactory->create($providerKey, $configuration);
-        $user = $client->getUser($state, $code, $behaviour);
+        $oauthClient = $this->clientFactory->create($client->provider, $client->config);
+        $user = $oauthClient->getUser($state, $code, $behaviour);
         $user->addArrayExtension('requestState', [
             'requestState' => $state,
         ]);
 
-        $this->discoverRoleAssignment($rules, $user, $client, $configuration);
+        $client->addExtension('oauthClient', $oauthClient);
 
-        $this->eventDispatcher->dispatch(new UserRedirectReceivedEvent($user, $request, $behaviour));
+        $this->eventDispatcher->dispatch(
+            new UserRedirectReceivedEvent(
+                $user,
+                $client,
+                $request,
+                $behaviour
+            )
+        );
 
         return $user;
-    }
-
-    private function discoverRoleAssignment(
-        ClientRuleCollection $rules,
-        User $user,
-        ClientContract $client,
-        array $configuration
-    ): void {
-        $ruleScope = new OAuthRuleScope($user, $client, $configuration, Context::createDefaultContext(), $this->logger);
-        $client->prepareOAuthRuleScope($ruleScope);
-
-        $roleAssignment = new RoleAssignment();
-        $user->addExtension('roleAssignment', $roleAssignment);
-
-        foreach ($rules->getElements() as $rule) {
-            if ($this->clientRuleValidator->isValid($rule->getId(), $ruleScope)) {
-                if (!$roleAssignment->isAdministrator) {
-                    $roleAssignment->isAdministrator = $rule->isUserBecomeAdmin();
-                }
-
-                $roleAssignment->roleIds = [
-                    ...$roleAssignment->roleIds,
-                    ...$rule->getAclRoles()->getIds(),
-                ];
-
-                if ($rule->isStopOnMatch()) {
-                    break;
-                }
-            }
-        }
     }
 }