+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Added**
 
 * Added extensible login actions to allow custom actions after a successful login
+* Added configuration to automatically redirect users to the identity provider for login
 
 **Changed**