+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Added**
 
 * Added extensible login actions to allow custom actions after a successful login
+* Added `bcmath` and `gmp` to suggested php extensions in `composer.json` (GitHub Issue #5)
 
 **Changed**